requirements.txt
Django==3.2.16
django-allauth
django-appconf
django-autoslug
django-ckeditor
django-compressor
django-js-asset
django-quill-editor
django-tinymce
idna
oauthlib
Pillow
pycodestyle
pycparser
PyJWT
python3-openid
pytz
rcssmin
requests
requests-oauthlib
rjsmin
sorl-thumbnail
sqlparse
toml
types-cryptography
urllib3
